py-doc-to-xml/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
click==8.0.4
    # via
    #   -r requirements.in
    #   typer
lxml==4.8.0
    # via
    #   -r requirements.in
    #   python-docx
    #   simplify-docx
more-itertools==7.0.0
    # via simplify-docx
pydantic==1.9.0
    # via -r requirements.in
python-docx==0.8.11
    # via simplify-docx
simplify-docx @ git+https://github.com/microsoft/Simplify-Docx.git
    # via -r requirements.in
six==1.16.0
    # via
    #   -r requirements.in
    #   simplify-docx
typer==0.4.1
    # via -r requirements.in
typing-extensions==4.1.1
    # via pydantic
wincertstore==0.2
    # via simplify-docx
